JS-4524: Move render methods
