+++ src/ts/component/block/index.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
+import * as ReactDOM from 'react-dom';
 import $ from 'jquery';
 import { observer } from 'mobx-react';
-import { I, C, S, U, J, keyboard, focus, Storage, Preview, Renderer, Mark } from 'Lib';
-import { DropTarget, ListChildren, Icon, SelectionTarget, IconObject } from 'Component';
+import { I, C, S, U, J, keyboard, focus, Storage, Preview, Renderer, Mark, translate } from 'Lib';
+import { DropTarget, ListChildren, Icon, SelectionTarget, IconObject, Loader } from 'Component';
 
 import BlockDataview from './dataview';
 import BlockText from './text';
@@ -65,6 +66,9 @@ const Block = observer(class Block extends React.Component<Props> {
 		this.onMouseLeave = this.onMouseLeave.bind(this);
 		this.onContextMenu = this.onContextMenu.bind(this);
 		this.renderLinks = this.renderLinks.bind(this);
+		this.renderMentions = this.renderMentions.bind(this);
+		this.renderObjects = this.renderObjects.bind(this);
+		this.renderEmoji = this.renderEmoji.bind(this);
 	};
 
 	render () {
@@ -132,7 +136,18 @@ const Block = observer(class Block extends React.Component<Props> {
 					canDrop = false;
 				};
 
-				blockComponent = <BlockText key={key} ref={setRef} {...this.props} onToggle={this.onToggle} renderLinks={this.renderLinks} />;
+				blockComponent = (
+					<BlockText 
+						key={key} 
+						ref={setRef} 
+						{...this.props} 
+						onToggle={this.onToggle} 
+						renderLinks={this.renderLinks} 
+						renderMentions={this.renderMentions}
+						renderObjects={this.renderObjects}
+						renderEmoji={this.renderEmoji}
+					/>
+				);
 				break;
 			};
 
@@ -208,7 +223,17 @@ const Block = observer(class Block extends React.Component<Props> {
 
 			case I.BlockType.Chat: {
 				canDrop = canSelect = !root.isObjectChat();
-				blockComponent = <BlockChat key={key} ref={setRef} {...this.props} renderLinks={this.renderLinks} />;
+				blockComponent = (
+					<BlockChat 
+						key={key} 
+						ref={setRef} 
+						{...this.props} 
+						renderLinks={this.renderLinks} 
+						renderMentions={this.renderMentions}
+						renderObjects={this.renderObjects}
+						renderEmoji={this.renderEmoji}
+					/>
+				);
 				break;
 			};
 				
@@ -762,10 +787,10 @@ const Block = observer(class Block extends React.Component<Props> {
 		});
 	};
 
-	renderLinks (node: any, marks: I.Mark[], canEdit: boolean, onChange: (marks: I.Mark[]) => void) {
+	renderLinks (node: any, marks: I.Mark[], value: string) {
 		node = $(node);
 
-		const { rootId } = this.props;
+		const { rootId, readonly } = this.props;
 		const items = node.find(Mark.getTag(I.MarkType.Link));
 
 		if (!items.length) {
@@ -815,9 +840,9 @@ const Block = observer(class Block extends React.Component<Props> {
 					to: Number(range[1]) || 0, 
 				},
 				marks,
-				onChange,
-				noUnlink: !canEdit,
-				noEdit: !canEdit,
+				onChange: marks => this.setMarks(value, marks),
+				noUnlink: readonly,
+				noEdit: readonly,
 			});
 
 			element.off('click.link').on('click.link', e => {
@@ -825,6 +850,229 @@ const Block = observer(class Block extends React.Component<Props> {
 				isInside ? U.Router.go(route, {}) : Renderer.send('urlOpen', target);
 			});
 		});
+	};
+
+	renderMentions (node: any, marks: I.Mark[], value: string) {
+		node = $(node);
+
+		const { rootId, block } = this.props;
+		const size = U.Data.emojiParam(block.content.style);
+		const items = node.find(Mark.getTag(I.MarkType.Mention));
+		
+		if (!items.length) {
+			return;
+		};
+
+		items.each((i: number, item: any) => {
+			item = $(item);
+			
+			const data = item.data();
+			if (!data.param) {
+				return;
+			};
+
+			const smile = item.find('smile');
+			if (!smile.length) {
+				return;
+			};
+
+			const object = S.Detail.get(rootId, data.param, []);
+			const { id, _empty_, layout, done, isDeleted, isArchived } = object;
+			const isTask = U.Object.isTaskLayout(layout);
+			const name = item.find('name');
+			const clickable = isTask ? item.find('name') : item;
+
+			let icon = null;
+			if (_empty_) {
+				icon = <Loader type="loader" className={[ 'c' + size, 'inline' ].join(' ')} />;
+			} else {
+				icon = (
+					<IconObject 
+						id={`mention-${block.id}-${i}`}
+						size={size} 
+						object={object} 
+						canEdit={!isArchived && isTask} 
+						onSelect={icon => this.onMentionSelect(value, marks, id, icon)} 
+						onUpload={objectId => this.onMentionUpload(value, marks, id, objectId)} 
+						onCheckbox={() => this.onMentionCheckbox(value, marks, id, !done)}
+					/>
+				);
+			};
+
+			if (_empty_ || isDeleted) {
+				item.addClass('disabled');
+			};
+
+			if ((layout == I.ObjectLayout.Task) && done) {
+				item.addClass('isDone');
+			};
+
+			ReactDOM.render(icon, smile.get(0), () => {
+				if (smile.html()) {
+					item.addClass('withImage c' + size);
+				};
+			});
+
+			clickable.off('mouseenter.mention');
+			clickable.on('mouseenter.mention', e => {
+				const sr = U.Common.getSelectionRange();
+				if (sr && !sr.collapsed) {
+					return;
+				};
+
+				const range = String(item.attr('data-range') || '').split('-');
+				const param = String(item.attr('data-param') || '');
+
+				if (!param || item.hasClass('disabled')) {
+					return;
+				};
+
+				const object = S.Detail.get(rootId, param, []);
+
+				clickable.off('click.mention').on('click.mention', e => {
+					e.preventDefault();
+					U.Object.openEvent(e, object);
+				});
+
+				Preview.previewShow({
+					target: object.id,
+					element: name,
+					range: { 
+						from: Number(range[0]) || 0,
+						to: Number(range[1]) || 0, 
+					},
+					noUnlink: true,
+					marks,
+					onChange: marks => this.setMarks(value, marks),
+				});
+			});
+		});
+	};
+
+	renderObjects (node: any, marks: I.Mark[], value: string) {
+		node = $(node);
+
+		const { rootId, readonly } = this.props;
+		const items = node.find(Mark.getTag(I.MarkType.Object));
+
+		if (!items.length) {
+			return;
+		};
+
+		items.each((i: number, item: any) => {
+			item = $(item);
+			
+			const param = item.attr('data-param');
+			const object = S.Detail.get(rootId, param, []);
+
+			if (object._empty_ || object.isDeleted) {
+				item.addClass('disabled');
+			};
+		});
+
+		items.off('mouseenter.object mouseleave.object');
+		items.on('mouseleave.object', () => Preview.tooltipHide(false));
+		items.on('mouseenter.object', e => {
+			const sr = U.Common.getSelectionRange();
+			if (sr && !sr.collapsed) {
+				return;
+			};
+
+			const element = $(e.currentTarget);
+			const range = String(element.attr('data-range') || '').split('-');
+			const param = String(element.attr('data-param') || '');
+			const object = S.Detail.get(rootId, param, []);
+			
+			let tt = '';
+			if (object.isDeleted) {
+				tt = translate('commonDeletedObject');
+			};
+
+			if (tt) {
+				Preview.tooltipShow({ text: tt, element });
+				return;
+			};
+
+			if (!param || object.isDeleted) {
+				return;
+			};
+
+			element.off('click.object').on('click.object', e => {
+				e.preventDefault();
+				U.Object.openEvent(e, object);
+			});
+
+			Preview.previewShow({
+				target: object.id,
+				object,
+				element,
+				marks,
+				onChange: marks => this.setMarks(value, marks),
+				range: { 
+					from: Number(range[0]) || 0,
+					to: Number(range[1]) || 0, 
+				},
+				noUnlink: readonly,
+				noEdit: readonly,
+			});
+		});
+	};
+
+	renderEmoji (node: any) {
+		node = $(node);
+
+		const items = node.find(Mark.getTag(I.MarkType.Emoji));
+		if (!items.length) {
+			return;
+		};
+
+		const { block } = this.props;
+		const size = U.Data.emojiParam(block.content.style);
+
+		items.each((i: number, item: any) => {
+			item = $(item);
+
+			const param = item.attr('data-param');
+			const smile = item.find('smile');
+
+			if (smile.length) {
+				ReactDOM.render(<IconObject size={size} object={{ iconEmoji: param }} />, smile.get(0));
+			};
+		});
+	};
+
+	onMentionSelect (value: string, marks: I.Mark[], id: string, icon: string) {
+		const { rootId, block } = this.props;
+
+		U.Data.blockSetText(rootId, block.id, value, marks, true, () => {
+			U.Object.setIcon(id, icon, '');
+		});
+	};
+
+	onMentionUpload (value: string, marks: I.Mark[], targetId: string, objectId: string) {
+		const { rootId, block } = this.props;
+
+		U.Data.blockSetText(rootId, block.id, value, marks, true, () => {
+			U.Object.setIcon(targetId, '', objectId);
+		});
+	};
+
+	onMentionCheckbox (value: string, marks: I.Mark[], objectId: string, done: boolean) {
+		const { rootId, block } = this.props;
+
+		U.Data.blockSetText(rootId, block.id, value, marks, true, () => {
+			U.Object.setDone(objectId, done);
+		});
+	};
+
+	setMarks (value: string, marks: I.Mark[]) {
+		const { rootId, block } = this.props;
+		
+		if (block.isTextCode()) {
+			marks = [];
+		};
+
+		U.Data.blockSetText(rootId, block.id, value, marks, true);
 	};
 
 });
