+++ src/ts/component/block/text.tsx
@@ -5,7 +5,7 @@ import $ from 'jquery';
 import raf from 'raf';
 import { observer, } from 'mobx-react';
 import { Select, Marker, Loader, IconObject, Icon, Editable } from 'Component';
-import { I, C, S, U, J, keyboard, Key, Preview, Mark, focus, Storage, translate, analytics, Renderer } from 'Lib';
+import { I, C, S, U, J, keyboard, Key, Preview, Mark, focus, Storage, translate, analytics } from 'Lib';
 
 interface Props extends I.BlockComponent {
 	onToggle?(e: any): void;
@@ -63,7 +63,6 @@ const BlockText = observer(class BlockText extends React.Component<Props> {
 		this.onCopy = this.onCopy.bind(this);
 		this.onSelectIcon = this.onSelectIcon.bind(this);
 		this.onUploadIcon = this.onUploadIcon.bind(this);
-		this.setMarks = this.setMarks.bind(this);
 	};
 
 	render () {
@@ -249,7 +248,7 @@ const BlockText = observer(class BlockText extends React.Component<Props> {
 	};
 
 	setValue (v: string) {
-		const { block, readonly, renderLinks } = this.props;
+		const { block, renderLinks, renderObjects, renderMentions, renderEmoji } = this.props;
 		const fields = block.fields || {};
 		
 		let text = String(v || '');
@@ -297,11 +296,10 @@ const BlockText = observer(class BlockText extends React.Component<Props> {
 			};
 
 			this.frame = raf(() => {
-				renderLinks(this.node, this.marks, !readonly, this.setMarks);
-
-				this.renderObjects();
-				this.renderMentions();
-				this.renderEmoji();
+				renderLinks(this.node, this.marks, html);
+				renderMentions(this.node, this.marks, html);
+				renderObjects(this.node, this.marks, html);
+				renderEmoji(this.node);
 			});
 		};
 
@@ -310,233 +308,6 @@ const BlockText = observer(class BlockText extends React.Component<Props> {
 		};
 	};
 	
-	renderObjects () {
-		if (!this._isMounted) {
-			return;
-		};
-
-		const { rootId, readonly } = this.props;
-		const node = $(this.node);
-		const items = node.find(Mark.getTag(I.MarkType.Object));
-
-		if (!items.length) {
-			return;
-		};
-
-		items.each((i: number, item: any) => {
-			item = $(item);
-			
-			const data = item.data();
-			if (!data.param) {
-				return;
-			};
-
-			const object = S.Detail.get(rootId, data.param, []);
-			const { _empty_, isDeleted } = object;
-
-			if (_empty_ || isDeleted) {
-				item.addClass('disabled');
-			};
-		});
-
-		items.off('mouseenter.object mouseleave.object');
-		items.on('mouseleave.object', () => Preview.tooltipHide(false));
-		items.on('mouseenter.object', e => {
-			const sr = U.Common.getSelectionRange();
-			if (sr && !sr.collapsed) {
-				return;
-			};
-
-			const element = $(e.currentTarget);
-			const range = String(element.attr('data-range') || '').split('-');
-			const param = String(element.attr('data-param') || '');
-			const object = S.Detail.get(rootId, param, []);
-			
-			let tt = '';
-			if (object.isDeleted) {
-				tt = translate('commonDeletedObject');
-			};
-
-			if (tt) {
-				Preview.tooltipShow({ text: tt, element });
-				return;
-			};
-
-			if (!param || object.isDeleted) {
-				return;
-			};
-
-			element.off('click.object').on('click.object', e => {
-				e.preventDefault();
-				U.Object.openEvent(e, object);
-			});
-
-			Preview.previewShow({
-				target: object.id,
-				object,
-				element,
-				range: { 
-					from: Number(range[0]) || 0,
-					to: Number(range[1]) || 0, 
-				},
-				marks: this.marks,
-				onChange: this.setMarks,
-				noUnlink: readonly,
-				noEdit: readonly,
-			});
-		});
-	};
-
-	renderMentions () {
-		if (!this._isMounted) {
-			return;
-		};
-
-		const node = $(this.node);
-		const items = node.find(Mark.getTag(I.MarkType.Mention));
-		
-		if (!items.length) {
-			return;
-		};
-
-		const { rootId, block } = this.props;
-		const size = this.emojiParam(block.content.style);
-
-		items.each((i: number, item: any) => {
-			item = $(item);
-			
-			const data = item.data();
-			if (!data.param) {
-				return;
-			};
-
-			const smile = item.find('smile');
-			if (!smile.length) {
-				return;
-			};
-
-			const object = S.Detail.get(rootId, data.param, []);
-			const { id, _empty_, layout, done, isDeleted, isArchived } = object;
-			const isTask = U.Object.isTaskLayout(layout);
-			const name = item.find('name');
-			const clickable = isTask ? item.find('name') : item;
-
-			let icon = null;
-			if (_empty_) {
-				icon = <Loader type="loader" className={[ 'c' + size, 'inline' ].join(' ')} />;
-			} else {
-				icon = (
-					<IconObject 
-						id={`mention-${block.id}-${i}`}
-						size={size} 
-						object={object} 
-						canEdit={!isArchived && isTask} 
-						onSelect={icon => this.onMentionSelect(id, icon)} 
-						onUpload={objectId => this.onMentionUpload(id, objectId)} 
-						onCheckbox={() => this.onMentionCheckbox(id, !done)}
-					/>
-				);
-			};
-
-			if (_empty_ || isDeleted) {
-				item.addClass('disabled');
-			};
-
-			if ((layout == I.ObjectLayout.Task) && done) {
-				item.addClass('isDone');
-			};
-
-			ReactDOM.render(icon, smile.get(0), () => {
-				if (smile.html()) {
-					item.addClass('withImage c' + size);
-				};
-			});
-
-			clickable.off('mouseenter.mention');
-			clickable.on('mouseenter.mention', e => {
-				const sr = U.Common.getSelectionRange();
-				if (sr && !sr.collapsed) {
-					return;
-				};
-
-				const range = String(item.attr('data-range') || '').split('-');
-				const param = String(item.attr('data-param') || '');
-
-				if (!param || item.hasClass('disabled')) {
-					return;
-				};
-
-				const object = S.Detail.get(rootId, param, []);
-
-				clickable.off('click.mention').on('click.mention', e => {
-					e.preventDefault();
-					U.Object.openEvent(e, object);
-				});
-
-				Preview.previewShow({
-					target: object.id,
-					element: name,
-					range: { 
-						from: Number(range[0]) || 0,
-						to: Number(range[1]) || 0, 
-					},
-					marks: this.marks,
-					noUnlink: true,
-					onChange: this.setMarks,
-				});
-			});
-		});
-	};
-
-	renderEmoji () {
-		if (!this._isMounted) {
-			return;
-		};
-
-		const node = $(this.node);
-		const items = node.find(Mark.getTag(I.MarkType.Emoji));
-		
-		if (!items.length) {
-			return;
-		};
-
-		const { block } = this.props;
-		const size = this.emojiParam(block.content.style);
-
-		items.each((i: number, item: any) => {
-			item = $(item);
-
-			const data = item.data();
-			if (!data.param) {
-				return;
-			};
-
-			const smile = item.find('smile');
-			if (smile.length) {
-				ReactDOM.render(<IconObject size={size} object={{ iconEmoji: data.param }} />, smile.get(0));
-			};
-		});
-	};
-
-	emojiParam (style: I.TextStyle) {
-		let size = 24;
-		switch (style) {
-			case I.TextStyle.Header1:
-				size = 32;
-				break;
-			
-			case I.TextStyle.Header2:
-				size = 28;
-				break;
-
-			case I.TextStyle.Header3:
-			case I.TextStyle.Quote:
-				size = 26;
-				break;
-		};
-		return size;
-	};
-
 	getValue (): string {
 		return this.refEditable ? this.refEditable.getTextValue() : '';
 	};
@@ -1158,17 +929,6 @@ const BlockText = observer(class BlockText extends React.Component<Props> {
 		U.Data.blockSetText(rootId, block.id, value, marks, update, callBack);
 	};
 	
-	setMarks (marks: I.Mark[]) {
-		const { rootId, block } = this.props;
-		const value = this.getValue();
-		
-		if (block.isTextCode()) {
-			marks = [];
-		};
-
-		U.Data.blockSetText(rootId, block.id, value, marks, true);
-	};
-	
 	onFocus (e: any) {
 		const { onFocus } = this.props;
 
@@ -1325,9 +1085,8 @@ const BlockText = observer(class BlockText extends React.Component<Props> {
 		this.timeoutContext = window.setTimeout(() => {
 			const onChange = (marks: I.Mark[]) => {
 				this.marks = marks;
-				this.setMarks(marks);
 
-				raf(() => {
+				U.Data.blockSetText(rootId, block.id, this.getValue(), this.marks, true, () => {
 					focus.set(block.id, { from: currentFrom, to: currentTo });
 					focus.apply();
 				});
@@ -1439,33 +1198,6 @@ const BlockText = observer(class BlockText extends React.Component<Props> {
 		if (this.refEditable) {
 			this.refEditable.placeholderHide();
 		};
-	};
-
-	onMentionSelect (id: string, icon: string) {
-		const { rootId, block } = this.props;
-		const value = this.getValue();
-
-		U.Data.blockSetText(rootId, block.id, value, this.marks, true, () => {
-			U.Object.setIcon(id, icon, '');
-		});
-	};
-
-	onMentionUpload (targetId: string, objectId: string) {
-		const { rootId, block } = this.props;
-		const value = this.getValue();
-
-		U.Data.blockSetText(rootId, block.id, value, this.marks, true, () => {
-			U.Object.setIcon(targetId, '', objectId);
-		});
-	};
-
-	onMentionCheckbox (objectId: string, done: boolean) {
-		const { rootId, block } = this.props;
-		const value = this.getValue();
-
-		U.Data.blockSetText(rootId, block.id, value, this.marks, true, () => {
-			U.Object.setDone(objectId, done);
-		});
 	};
 
 	checkMarkOnBackspace (value: string) {
